Move clique resizing functions to clique module
